+++ src/Plugin.php
@@ -8,6 +8,7 @@ use craft\base\Model;
 use craft\base\Plugin as BasePlugin;
 use craft\elements\Asset;
 use craft\events\DefineAssetThumbUrlEvent;
+use craft\events\DefineElementInnerHtmlEvent;
 use craft\events\DefineGqlTypeFieldsEvent;
 use craft\events\RegisterElementHtmlAttributesEvent;
 use craft\events\RegisterElementTableAttributesEvent;
@@ -15,6 +16,8 @@ use craft\events\RegisterGqlTypesEvent;
 use craft\events\SetElementTableAttributeHtmlEvent;
 use craft\events\TemplateEvent;
 use craft\gql\TypeManager;
+use craft\helpers\Cp;
+use craft\helpers\Html;
 use craft\helpers\Json;
 use craft\helpers\UrlHelper;
 use craft\services\Assets;
@@ -98,6 +101,7 @@ class Plugin extends BasePlugin
             $this->_registerPreviewHandler();
             $this->_registerSaveListener();
             $this->_registerDeleteListener();
+            $this->_registerLink();
 
             // if showThumbnailsInCp is set to true add in the asset index attribute for the thumbnails
             if ($this->getSettings()->showThumbnailsInCp) {
@@ -261,6 +265,31 @@ class Plugin extends BasePlugin
                 Craft::$app->getCache()->delete($this->methods->getCachedAssetKey($event->sender));
             }
         });
+    }
+
+    /**
+     * Adds link icons to embedded assets in asset fields.
+     */
+    private function _registerLink(): void
+    {
+        Event::on(
+            Cp::class,
+            Cp::EVENT_DEFINE_ELEMENT_INNER_HTML,
+            function(DefineElementInnerHtmlEvent $event) {
+                if (
+                    !$event->element instanceof Asset ||
+                    ($embeddedAsset = $this->methods->getEmbeddedAsset($event->element)) === null
+                ) {
+                    return;
+                }
+
+                $event->innerHtml .= Html::tag('a', '', [
+                    'class' => 'icon link',
+                    'href' => $embeddedAsset->url,
+                    'title' => Craft::t('embeddedassets', 'View'),
+                ]);
+            }
+        );
     }
 
     /**